Cambios

+++ src/Predicciones/prediccionSimple.cc
@@ -5,7 +5,8 @@ double SimplePrediction(std::vector<std::vector<double>> data, int item, std::ve
   double result, sum_of_items = 0, sum_of_similarities = 0;
 
   for (int i = 0; i < similarity_neighbors.size(); i++) {
-    int aux = (similarity_neighbors[i].second * data[similarity_neighbors[i].first][item]);
+    double aux = (similarity_neighbors[i].second * data[similarity_neighbors[i].first][item]);
+    // std::cout << aux << std::endl;
     sum_of_items += aux;
     sum_of_similarities += fabs(similarity_neighbors[i].second);
   }
